Added pretty_value method to transaction.rb to return formatted money strings. Modified DB to store values as DECIMAL, renamed DB to scunner. Added images for edit/delete transaction buttons. Added delete mechanism to transactions/index.erb, transactions/search_by_tag.erb and transaction controller. Minor CSS formatting nudges.

+++ db/money_cash_board.sql
@@ -16,7 +16,7 @@ CREATE TABLE tags (
 CREATE TABLE transactions (
   id SERIAL8 primary key,
   merchant_name VARCHAR(255),
-  value INT8,
+  value DECIMAL,
   person_id INT8 references people(id),
   tag_id INT8 references tags(id)
 );
